Update stark.md
It seems that ∃ is the quantifier that was meant here. I.e. for any value that a register took, there must be a combination in B that permits it:

- ∀ (i, w, e). (W[i, w] = e) ⇒ ((i, w, e) ∈ B)
- ∃ (i, w, e) ∈ B. (W[i, w] = e)

The above two are equivalent logical statements, either is correct. But I don't see a way of reading what was there before that makes sense.